[RISCV64] Add type coercions for type mismatches
Due to the introduction of using i64 for pointers, the discovered types
may not match a parameter type of a function signature. In these cases,
an i64 will be converted to a ptr using the IntToPtr instruction and a
ptr value will be converted to i64 using the PtrToInt instruction where
necessary.

In the case of integer values, there can be a mismatch in bit width of
the two values of a binary operation or comparison. At that point, the
smaller of the two values (in terms of width) will be sign-extended to
the type of the bigger of the two.